make sure the lag calculation is not truncated

+++ src/source.c
@@ -668,7 +668,7 @@ static int source_queue_advance (client_t *client)
 {
     source_t *source = client->shared_data;
     refbuf_t *refbuf;
-    int lag;
+    uint64_t lag;
 
     if (client->refbuf == NULL && locate_start_on_queue (source, client) < 0)
         return -1;
